media_processing_pipeline/requirements.txt
# A
aiohttp
# D
django-background-tasks

# F
ffmpeg-python
# G
google-cloud-speech 